Reorganize `filenameToLanguage`
